Premaster

+++ event-management-engine/src/main/java/com/webwork/event/management/service/impl/UserServiceImpl.java
@@ -3,18 +3,26 @@ package com.webwork.event.management.service.impl;
 import java.util.List;
 import java.util.Optional;
 
+import javax.mail.MessagingException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.webwork.event.management.email.AbstractEmailContext;
+import com.webwork.event.management.email.AccountDeletedEmailContext;
 import com.webwork.event.management.entity.User;
 import com.webwork.event.management.exception.EntityNotFoundException;
 import com.webwork.event.management.repository.UserRepository;
+import com.webwork.event.management.service.EmailService;
 import com.webwork.event.management.service.UserService;
 
 @Service
 public class UserServiceImpl implements UserService {
 
+	@Autowired
+	private EmailService emailService;
+	
 	@Autowired
 	private UserRepository userRepo;
 
@@ -46,15 +54,26 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	@Transactional
-	public boolean delete(int id) {
+	public boolean delete(int id) throws MessagingException {
 		Optional<User> result = userRepo.findById(id);
 		if (!result.isPresent()) {
 			throw new EntityNotFoundException("User Not Found");
 		} else {
 			User user = result.get();
-			userRepo.delete(user);
-			return true;
+			if(sendEmail(user)) {
+				userRepo.delete(user);
+				return true;
+			}
+			
 		}
+		return false;
+	}
+
+	private boolean sendEmail(User user) throws MessagingException {
+		AbstractEmailContext emailContext = new AccountDeletedEmailContext();
+		emailContext.init(user);
+		emailService.sendMail(emailContext);
+		return true;
 	}
 
 }
